requirements.txt
# 报纸大样PDF解析系统依赖
# Newspaper Layout PDF Parser Requirements

# PDF处理
pymupdf>=1.23.0

# 科学计算
numpy>=1.24.0

# 机器学习（后续里程碑使用）
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.7.0

# 配置文件处理
PyYAML>=6.0
